pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "baby-sleep-scheduler"
version = "0.1.0"
description = "A terminal app that predicts baby sleep schedules using pattern-based ML"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "numpy>=1.24",
    "google-api-python-client>=2.100",
    "google-auth-httplib2>=0.1",
    "google-auth-oauthlib>=1.0",
]

[project.scripts]
baby-sleep = "baby_sleep.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]
